+++ src/main/java/com/CookingMama/dev/security/AdminTokenInfo.java
@@ -6,11 +6,11 @@ import lombok.*;
 @Getter@Setter@ToString
 @AllArgsConstructor@NoArgsConstructor
 public class AdminTokenInfo {
-    private Integer id;
+    private Long id;
     private String adminEmail;
     private String adminName;
     public AdminTokenInfo tokenToDTO(Claims claims){
-        Integer id = (Integer) claims.get("id");
+        Long id = (Long) claims.get("id");
         String adminEmail = (String) claims.get("adminEmail");
         String adminName = (String) claims.get("adminName");
         return new AdminTokenInfo(id, adminEmail, adminName);
